# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/systemmembers": {
            "get": {
                "description": "list systemmember entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List systemmember entities",
                "operationId": "list-systemmember",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Systemmember"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create ussystemmemberer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create systemmember",
                "operationId": "create-systemmember",
                "parameters": [
                    {
                        "description": "Systemmember entity",
                        "name": "systemmember",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Systemmember"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Systemmember"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/systemmembers/{id}": {
            "get": {
                "description": "get systemmember by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a systemmember entity by ID",
                "operationId": "get-systemmember",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Systemmember ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Systemmember"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update systemmember by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a systemmember entity by ID",
                "operationId": "update-systemmember",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Systemmember ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Systemmember entity",
                        "name": "systemmember",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Systemmember"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Systemmember"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get systemmember by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a systemmember entity by ID",
                "operationId": "delete-systemmember",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Systemmember ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "ent.Doctor": {
            "type": "object",
            "properties": {
                "Doctor_Email": {
                    "description": "DoctorEmail holds the value of the \"Doctor_Email\" field.",
                    "type": "string"
                },
                "Doctor_Name": {
                    "description": "DoctorName holds the value of the \"Doctor_Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DoctorQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DoctorEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.DoctorEdges": {
            "type": "object",
            "properties": {
                "prescriptions": {
                    "description": "Prescriptions holds the value of the prescriptions edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Prescription"
                    }
                }
            }
        },
        "ent.Medicine": {
            "type": "object",
            "properties": {
                "Medicine_Name": {
                    "description": "MedicineName holds the value of the \"Medicine_Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the MedicineQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MedicineEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.MedicineEdges": {
            "type": "object",
            "properties": {
                "prescriptions": {
                    "description": "Prescriptions holds the value of the prescriptions edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Prescription"
                    }
                }
            }
        },
        "ent.Patient": {
            "type": "object",
            "properties": {
                "Gender": {
                    "description": "Gender holds the value of the \"Gender\" field.",
                    "type": "string"
                },
                "Patient_Name": {
                    "description": "PatientName holds the value of the \"Patient_Name\" field.",
                    "type": "string"
                },
                "Patient_Phone": {
                    "description": "PatientPhone holds the value of the \"Patient_Phone\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PatientQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.PatientEdges": {
            "type": "object",
            "properties": {
                "prescriptions": {
                    "description": "Prescriptions holds the value of the prescriptions edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Prescription"
                    }
                }
            }
        },
        "ent.Prescription": {
            "type": "object",
            "properties": {
                "Datetime": {
                    "description": "Datetime holds the value of the \"Datetime\" field.",
                    "type": "string"
                },
                "doctor_ID": {
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PrescriptionQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PrescriptionEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "patient_ID": {
                    "type": "integer"
                },
                "systemmember_ID": {
                    "type": "integer"
                }
            }
        },
        "ent.PrescriptionEdges": {
            "type": "object",
            "properties": {
                "doctor": {
                    "description": "Doctor holds the value of the doctor edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctor"
                },
                "medicines": {
                    "description": "Medicines holds the value of the medicines edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Medicine"
                    }
                },
                "patient": {
                    "description": "Patient holds the value of the patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Patient"
                },
                "systemmember": {
                    "description": "Systemmember holds the value of the systemmember edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Systemmember"
                }
            }
        },
        "ent.Systemmember": {
            "type": "object",
            "properties": {
                "Password": {
                    "description": "Password holds the value of the \"Password\" field.",
                    "type": "string"
                },
                "Systemmember_Name": {
                    "description": "SystemmemberName holds the value of the \"Systemmember_Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the SystemmemberQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SystemmemberEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.SystemmemberEdges": {
            "type": "object",
            "properties": {
                "prescriptions": {
                    "description": "Prescriptions holds the value of the prescriptions edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Prescription"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}